% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2D.R
\name{anomaly}
\alias{anomaly}
\title{Simple anomaly filter}
\usage{
anomaly(data, fun = "median", replace = FALSE, v = "SvANomal")
}
\arguments{
\item{data}{data containing the x, y and dat avalues}

\item{fun}{string describing the funciton to be performed, defaults to 'mean', other examples would be meidan, max, min, sd, var,...}
}
\description{
Computes a mean or median or whatever funciton is selected on a per depth bin basis for the entire mission (in the linear domain, back-transformed into log-space), which will then be subtracted from each value point, to compute an anomaly matrix. Additionally these values will then be scaled to fall within -1 and 1 for easier interpretation.
}
