% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvm_app.R
\name{dvm_app}
\alias{dvm_app}
\title{Shiny App to detect DVM}
\usage{
dvm_app(path)
}
\arguments{
\item{path}{Path to the ZOnar csv files}
}
\value{
shiny app
}
\description{
Shiny App to detect DVM
}
\examples{
path='Z:/zonar_sven/zonar/data/csv_zonar'
dvm_app(path)
}
\author{
Sven Gastauer
}
