% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env.R
\name{env_for_ac}
\alias{env_for_ac}
\title{Joinf acoustic and environmental data}
\usage{
env_for_ac(sv, env)
}
\arguments{
\item{sv}{Sv dataframe}

\item{env}{ENvironment dataframe}
}
\value{
combined dataframe
}
\description{
Joinf acoustic and environmental data
}
\author{
Sven Gastauer
}
