% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2D.R
\name{filter2d}
\alias{filter2d}
\title{Applies a 2D filter to the data}
\usage{
filter2d(
  data,
  x = 3,
  y = 3,
  xval = "Dive",
  yval = "Depth_r",
  val = "Sv",
  log = TRUE,
  fun = "mean"
)
}
\arguments{
\item{data}{data containing the x, y and dat avalues}

\item{x}{window size in x direction, defaults 3}

\item{y}{window size in y direction, defaults 3}

\item{xval}{the variable in data that describes the x axis, defaults 'Dive'}

\item{yval}{the varibale in data that describes the y axis, defaults 'Depth_r'}

\item{val}{the variable that describes the value variable in data, defaults 'Sv'}

\item{log}{if TRUE the value variable will be linearised andbefore the running calculation and backtransformed before being returned, defaults TRUE}

\item{fun}{string describing the funciton to be performed, defaults to 'mean', other examples would be meidan, max, min, sd, var,...}
}
\description{
Applies a 2D filter to the data
}
