% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accal.R
\name{get_accal}
\alias{get_accal}
\title{Get acoustic calibration information from netCDF file}
\usage{
get_accal(ncpath, mission)
}
\arguments{
\item{ncpath}{folder that contains the netCDF files}

\item{mission}{name or index number of the mission to be used}
}
\value{
dataframe containing  acoustic calibraiton information
}
\description{
Get acoustic calibration information from netCDF file
}
\author{
Sven Gastauer
}
