% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daydiff.R
\name{get_daydiff}
\alias{get_daydiff}
\title{Computes the difference between neighbouring day or night data, night for day and day for night}
\usage{
get_daydiff(
  data,
  fun = "median",
  days_before = 1,
  days_after = 1,
  replace = FALSE
)
}
\arguments{
\item{data}{Sv data}

\item{fun}{character either mean or median, defualts to median}

\item{days_before}{How many days before the given data point to include, defaults 1}

\item{days_after}{How many days after the given data point to include, defaults 1}

\item{replace}{Boolean, replaces Sv by diff or not, default FALSE}
}
\description{
Computes the difference between neighbouring day or night data, night for day and day for night
}
