% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{get_dvm}
\alias{get_dvm}
\title{Function to get the estimated depth of the DVM based on Sv values and a cutoff percentage}
\usage{
get_dvm(sv, cutval)
}
\arguments{
\item{sv}{dataframe containing Sv and depth information for each dive}

\item{cutval}{float, describing the cumulative percentage at which the cut should be made}
}
\value{
dataframe contains the per dive dvm depth
}
\description{
Function to get the estimated depth of the DVM based on Sv values and a cutoff percentage
}
\author{
Sven Gastauer
}
