% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env.R
\name{get_env}
\alias{get_env}
\title{Get environmental data from netCDF file}
\usage{
get_env(ncpath, mission)
}
\arguments{
\item{ncpath}{folder that contains the netCDF files}

\item{mission}{name or index number of the mission to be used}
}
\value{
dataframe contianing the temperature, salinity and fluorescence data per dive and depth interval
}
\description{
Currently reads in Temperature, Salinity and fluorescence information as recorded by the Zooglider CTD
}
\author{
Sven Gastauer
}
