% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{get_gps}
\alias{get_gps}
\title{Gets GPS and daytime for a given misson per dive}
\usage{
get_gps(ncpath, mission)
}
\arguments{
\item{ncpath}{path to the ZOnar netCDF files}

\item{mission}{string or integer which is the name of the mission to be loaded}
}
\value{
dataframe containing per dive mean, start and end lon, lat, and time for each dive
}
\description{
Gets GPS and daytime for a given misson per dive
}
\author{
Sven Gastauer
}
