% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{get_mission}
\alias{get_mission}
\title{get available files and missions in a given folder}
\usage{
get_mission(path, patterns = c("200kHz", "1000kHz", "Delta"), ending = "*.csv")
}
\arguments{
\item{path}{path that contains the zonar sv files}

\item{patterns}{patterns considered, defaults to c('200kHz', '1000kHz','Delta')}
}
\value{
list contiang the paths to the files for each pattern and a list of detected missions
}
\description{
get available files and missions in a given folder
}
