% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roi.R
\name{get_roi_counts}
\alias{get_roi_counts}
\title{Gets the ROI counts for a given mission}
\usage{
get_roi_counts(roidir, mission)
}
\arguments{
\item{roidir}{path to the csv files contianing the ROI informaiton}

\item{mission}{numeric or character, pointing to the filename to be considered}
}
\description{
Gets the ROI counts for a given mission
}
