% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{get_sv}
\alias{get_sv}
\title{Read the processed Zonar files
This function reads all acoustic Sv data for a given mission
It takes a path where the csv files containing the Sv informaiton are located and a mission reference.
The mission reference can be given as a number or character}
\usage{
get_sv(path, mission, ncpath = NULL)
}
\arguments{
\item{path}{character path to where the csv files are located}

\item{mission}{character or integer if a character is provided, this has to be the name of the mission as it is contained in the csv files name, if a number is provded the ith mission contained in the path will be used}

\item{ncpath}{path to the nc files, defaults to NULL. If NULL the parent folder of the csv path is assumed to contain a folder called nc_zonar, which contains the ZOnar netCDF files}
}
\description{
Read the processed Zonar files
This function reads all acoustic Sv data for a given mission
It takes a path where the csv files containing the Sv informaiton are located and a mission reference.
The mission reference can be given as a number or character
}
\author{
Sven Gastauer
}
