% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roi.R
\name{join_ac_roi}
\alias{join_ac_roi}
\title{Joins acoustic Sv data with ROI counts}
\usage{
join_ac_roi(Sv, rois)
}
\arguments{
\item{Sv}{Sv data list}

\item{rois}{rois dataframe}
}
\value{
combined dataframe
}
\description{
Joins acoustic Sv data with ROI counts
}
