% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{ncpath_from_csvpath}
\alias{ncpath_from_csvpath}
\title{Gets the standard nc path from a given csv path}
\usage{
ncpath_from_csvpath(csvpath)
}
\arguments{
\item{csvpath}{path to the ZOnar csv files}
}
\value{
path to nc files
}
\description{
This assumes that the standard file structure is kept intact
}
\author{
Sven Gastauer
}
