% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{plot_gps}
\alias{plot_gps}
\title{Plot mission map with bathymetry}
\usage{
plot_gps(gps, bathy = TRUE, daynight = TRUE, startend = TRUE)
}
\arguments{
\item{gps}{dataframe contianing grps information, generated by get_gps}

\item{bathy}{TRUE/FALSE to add bathymetry from marmap package or not, defaults to TRUE}

\item{daynight}{TRUE/FALSE color code points by day, night or dusk/dawn, defaults to TRUE}

\item{startend}{TRUE/FALSE shape points for start and end points of dive, defaults to TRUE}
}
\value{
ggplot
}
\description{
Plot mission map with bathymetry
}
\author{
Sven Gastauer
}
