% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roi.R
\name{plot_roi}
\alias{plot_roi}
\title{Plots ROI data in a grid}
\usage{
plot_roi(
  rois,
  roisel = NULL,
  skipdive = NULL,
  dres = 1,
  vmin = NULL,
  vmax = NULL
)
}
\arguments{
\item{rois}{dataframe contianing the roi information}

\item{roisel}{names of rois to be considered, NULL keeps all ROI classes, can be for example c('ROI25', 'ROI75','ROI200')}

\item{skipdive}{dive for which data should be excluded, NULL keeps all dives, could be for example c(41,43)}

\item{dres}{Depth resolution in m, default is 1}
}
\value{
ROI plots
}
\description{
This assumes that the standard file structure is kept intact
}
\author{
Sven Gastauer
}
