% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{plot_sv}
\alias{plot_sv}
\title{Plot sv data with day/night bar}
\usage{
plot_sv(
  d1,
  gps = NULL,
  cols = c(Night = "black", `Dusk/Dawn` = "gray", Day = "yellow"),
  cmaps = c("RdYlBu", "RdBu"),
  svmin = -85,
  svmax = -45,
  deltamin = -10,
  deltamax = 10,
  variable = NULL,
  nam = NULL
)
}
\arguments{
\item{d1}{dataframe contians the sv, dpeth and dive information}

\item{gps}{dataframe containing lat, lon, and time information}

\item{cmaps}{colour maps for Sv data and Sv delta if needed, defaults to c("RdYlBu",'RdBu')}

\item{svmin}{minimum value for Sv color, defaults to -85}

\item{svmax}{maximum value for Sv color, defaults to -45}

\item{deltamin}{minimum value for delta Sv, defaults to -10}

\item{deltamax}{maximum value for delta Sv, defaults to +10}

\item{nam}{character for legend title}

\item{col}{color scheme for the day/night bar defaults to "Night" = "black", "Dusk/Dawn" = "gray", "Day" = "yellow"}
}
\value{
list of plots
}
\description{
Plot sv data with day/night bar
}
\author{
Sven Gastauer
}
