% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Sv data and gps for plotting with plot_sv}
\usage{
prepare_data(data, gps)
}
\arguments{
\item{data}{Sv dataframe}

\item{gps}{dataframe from netCDF Sat GPS data}
}
\value{
dataframe containing the information needed for plotting with plot_sv
}
\description{
Prepare Sv data and gps for plotting with plot_sv
}
