% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roi.R
\name{roipath_from_csvpath}
\alias{roipath_from_csvpath}
\title{Gets the standard ZC ROI path from a given csv path}
\usage{
roipath_from_csvpath(csvpath)
}
\arguments{
\item{csvpath}{path to the ZOnar csv files}
}
\value{
path to ROI csv files
}
\description{
This assumes that the standard file structure is kept intact
}
\author{
Sven Gastauer
}
