% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermocline.R
\name{thermocline}
\alias{thermocline}
\title{Get thermocline or another *cline}
\usage{
thermocline(
  temps,
  depths,
  tmax = NULL,
  tmin = NULL,
  pac = 0.05,
  r = 1,
  cw = "c",
  dmax = NULL,
  dmin = NULL
)
}
\description{
Get thermocline or another *cline
}
\author{
Sven Gastauer
}
