% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermocline.R
\name{thermocline}
\alias{thermocline}
\title{Get thermocline or another *cline}
\usage{
thermocline(
  temps,
  depths,
  tmax = NULL,
  tmin = NULL,
  pac = 0.05,
  r = 1,
  cw = "c",
  dmax = NULL,
  dmin = NULL
)
}
\arguments{
\item{temps}{array of temperatures, must be the same length ans depths}

\item{depths}{array of depths, must be the same length as temps}

\item{tmax}{set a maximum value to be considered, all values above this threshold will be set to tmax, NULL for no maximum, defaults NULL}

\item{tmin}{set a minimum value to be considered, all values below this threshold will be set to tmin, NULL for no minimum, defaults NULL}

\item{pac}{proportion to be considered a thermocline around the maximum slope detected, defaults 5% = 0.05}

\item{r}{defines the window size of rolling median to be computed, this is handy if there are a few single outliers, in the dataset, defaults to 1 which is a rolling median with a window size of 1, which equals all raw values}

\item{cw}{defines if positive, negative or both slopes should be considered, cw = both directions, c=from warm to cold (large to low value), w = from cold to warm only (low to large value), defaults c}

\item{dmax}{defines a maximum depth to be considered, greater depths will be ignored, defaults NULL, no maximum depth}

\item{dmin}{defines a minimum depth to be considered, shallower depths will be ignored, defaults NULL, no minimum depth}
}
\description{
Computes slopes (m_x) between each input value pair as
\deqn{ m_x = \frac{temp_{x+1} - temp_x}{depth_{x+1} - depth_x}}{%
m_x = (temp_[x+1] - temp_x) / (depth_[x+1] - depth_x)}
The maximum slope (max([m_1,...m_n])) is assumed to represent a thermocline. \cr
Slopes in positive or negative direction or in a single direction can be considered.\cr
A threshold range around the maximum slope can be defined. \cr
Rolling median values can be used to minimze the influence from few outliers.\cr
Maximum or minimum values as well as depths to be considered can be defined
}
\author{
Sven Gastauer
}
